backend/requirements.txt
Flask
pillow
pytesseract
flask-cors
